Initial implementation
Currently, this implements a minimal set of functions needed to operate
a ploop: creation, mounting/unmounting, resizing, snapshotting, getting
some information, as well as logging and error reporting functions.

There is also a somewhat rudimentary test suite, barely checking that
the basic functionality is working. It has been tested on both OpenVZ
and VZ7.

The missing pieces (compared to C ploop library) are:
* discarding (compacting)
* online copy
* replacing
* converting
* creating temporary snapshots
* some auxiliary stuff.

All that is to be added when needed.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>