Replace(): implement
This implements the wrapper around ploop_replace_image(),
together with tests for both online and offline replace.

Replace() is to be used to replace an (read-only, i.e.
not top level) ploop delta image by a different (but identical)
file. This is useful, for example, if the image was on NFS
and you want to replace it with a local copy.

Before doing a replace, some checks are performed. In particular,
the new image file should be clean and have the same size as the
old one. Note that the two images should be byte-by-byte identical,
although this is not checked by the library as such check
is pretty expensive.

Image to be replaced (a.k.a. current image) can be specified
by either its uuid, current file name, or level (in running ploop).

Offline replace is also supported and it basically means
that the image file name is replaced.

In case keepName flag is specified, after a successful
replacement the new file is renamed to have the same name
as the old one (thus no modification to DiskDescriptor.xml
is necessary). In this case the old file is deleted.

Now, offline replace with keepName flag is essentially
the same as "mv newfile curfile" (with some added checks
that the files appear to be the same etc).

Signed-off-by: Kir Kolyshkin <kir@openvz.org>