test: introduce a way to abort tests
The tests here are not autonomous, they sometimes depend on a previous
tests. For example, if we can't create ploop image, we can't do any
other tests.

Because of this, there should be a way to abort running the tests.
As I failed to find a standard way, I mocked up my own abort()
function that:
 1 reports the error
 2 runs cleanup
 3 aborts the execution

Note that I had to use a direct write to stdout because logging
functions of testing package don't work if you do os.Exit() right
after. I suspect this is either buffering or paralellism.

Yes I know it looks ugly, but it seems to me it's less ugly to add
a check to every Test* function.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>