Use correct libs for static and dynamic link
Go (well, cgo) supports pkg-config directive so it's easier
to figure out which libs we need to link against. Note that
in case of static linking the list of libraries is larger
(as we also need to include the libraries needed by libraries
we use). Fortunately, there is --static flag to pkg-config to
aim in this.

Unfortunately, cgo doesn't know whether to pass this flag or not,
neither do we, so we have to ask the user to indicate they are
building statically by providing "static" build flag.

This solves the problem of "overlinking" the non-static binary.

Before:
 $ objdump -p goploop.test | grep NEEDED
  NEEDED               libploop.so.1
  NEEDED               librt.so.1
  NEEDED               libpthread.so.0
  NEEDED               libxml2.so.2
  NEEDED               libz.so.1
  NEEDED               liblzma.so.5
  NEEDED               libm.so.6
  NEEDED               libc.so.6

After:
 $ objdump -p goploop.test | grep NEEDED
  NEEDED               libploop.so.1
  NEEDED               libpthread.so.0
  NEEDED               libc.so.6

So, the libraries that are used by libploop are not explicitly
listed now.

For a discussion leading to this change, please see
https://github.com/golang/go/issues/12058

Signed-off-by: Kir Kolyshkin <kir@openvz.org>